+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Add Pastanaga Icon System @sneridagh
 * Support for nested schemas @robgietema
+* New block on return in editor @robgietema
 
 ### Changes